+++ app/controllers/jewels_controller.rb
@@ -2,10 +2,12 @@ class JewelsController < ApplicationController
   before_action :load_categories, only: ["new", "edit"]
   
   def new
+    redirect_to root_path if !defined?(current_user) || current_user.regular?
     @jewel = Jewel.new()
   end
   
   def create
+    redirect_to root_path if !defined?(current_user) || current_user.regular?
     params.permit!
     @jewel = Jewel.new(params[:jewel])
     @jewel.save
@@ -13,14 +15,16 @@ class JewelsController < ApplicationController
   end
   
   def edit
+    redirect_to root_path if !defined?(current_user) || current_user.regular?
     @jewel = Jewel.find_by_id(params[:id])    
   end
   
   def update
+    redirect_to root_path if !defined?(current_user) || current_user.regular?
     params.permit!
     @jewel = Jewel.find_by_id(params[:id])
     @jewel.update(params[:jewel])
-    redirect_to root_path
+    redirect_to jewels_path
   end
   
   def index
@@ -28,9 +32,10 @@ class JewelsController < ApplicationController
   end
   
   def destroy
+    redirect_to root_path if !defined?(current_user) || current_user.regular?
     jewel = Jewel.find_by_id(params[:id])
     jewel.destroy
-    redirect_to root_path
+    redirect_to jewels_path
   end
   
   private
